messy checkpoint: typechecking args and such
This is a half-baked implementation of hoisting. I gave up halfway in
favor of just leaning on class re-opening and manual forward
declaration. I want to figure this out eventually, since we'll need it
for Go-style package-level loading, but for now it's a bit of a time
sink.

Nonetheless it is partly brought back, and I'm committing this as-is
since git.dash seems to fully typecheck now.

Note: this raised the point that 0-arity (i.e. 0-params) functions
should be automatically called, which is already The Plan, and now in
hindsight I'm seeing it everywhere: self() becomes just self,
Directory() becomes Directory (I suppose?), etc.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>