parser: fix expr separators
Previously the PEG just lazily checked for an optional CommaToken
everywhere, which meant you could actually just place expressions
adjacent to one another and it'd be fine. This caused ambiguities
detected by Tree-Sitter which correctly complained that 'foo(bar)' was
ambiguous with 'foo (bar)'.

Now we explicitly check for either a linebreak or a comma.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>