initial support for reopening (<<)
Initially I wanted to support this form:

  foo { bar { baz = 2 } }

Which it turns out PureScript also implements. However, this syntax is
too ambiguous, e.g.:

  if foo { bar } else { baz }

So, I went with an explicit operator instead. I was tempted by <- but
not sure if I want to claim that operator yet.

Also, removed all the Record stuff which is unused; it's all Module now.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>