required args with defaults are exposed as nullable
This is the only real way for it to make sense, I think. If there's a
default, we shouldn't have to decide it on every call site - the
function should be able to determine it. Passing `null` is how to do
that.

Within the function body, we will keep the type non-null as declared.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>