fix dagger sdk for class constructors
a bit more verbose than i'd like, but working.

* fix ClassDecl Hoist setting incorrect type

Now we create the constructor type at hoisting time and set it, which we
need to, since that's what it is now. Also now we re-use FnDecl.Infer to
get the constructor type.

* switch back to just evaluating class bodies like before

The filtering was making things confusing. Instead, slot declarations
now just check if they already have a LOCALYL DEFINED value and no-op,
to handle constructor-supplied slots.

* fix public fields not being included when listing private bindings

Previously we checked for exact visibility matching, now we do `>=`. I
made private 0 instead of public which seems like a safer default.

* MarshalJSON now includes private fields

Should maybe be configurable, but the idea here is to support saving
private state in Dagger modules.

* add CompositeEnv.MarshalJSON

The last item ended up being the root cause for this, but keeping this
implementation anyway.

* adjust tests for class constructors

Now that public slots are converted into arguments for constructing
the class, we inherently don't have a `self` anymore. This makes sense,
because the entire point of the constructors is to prevent partial
instantiation. So, the tests were wrong. They should just refer to slots
without self.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>