re-use triple-quoted strings for multiline
I still like that Zig's syntax makes it extremely clear what the file
contents will be in the presence of linebreaks, but it's hard to justify
an additional syntax when our GraphQL lineage already prescribes
triple-quoted strings. We can just allow them anywhere, and assign the
same semantics we want for doc strings, considering a predominant use
case may be embedding prompts.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>