fix copy-on-write semantics w/ implicit self calls
The core issue is that when a method calls a sibling method without
explicit `self.`, what it's actually calling is the lexically-bound
`FunctionValue`, as opposed to a `BoundMethod` for `self.foo`. The
`FunctionValue` inside a class body actually has a statically bound
`self` slot, so when it mutates it, those mutations stick around for
future calls to the same FunctionValue.

To fix this:

* There is now a dynamically bound scope, accessed by the new keyword
  `self`.
* FunDecls now detect when they are defined within a dynamic scope (i.e.
  a class body), and FunctionValues bind themselves to the dynamic scope
  of the caller.
* BoundMethods set a fork of the receiver as their dynamic scope.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>